+++ src/mgp.h
@@ -75,23 +75,23 @@ struct mg_reader2 {
 
 /* ----------------------- geometry IO extern function ---------------------- */
 
-struct mg_object *mg_read_wkt(const char *wkt, int len);
-struct mg_object *mg_read_wkb(const char *wkb, int len);
-struct mg_object *mg_read_ewkt(const char *ewkt, int len);
-struct mg_object *mg_read_ewkb(const char *ewkb, int len);
-struct mg_object *mg_read_geojson(const char *json, int len);
-struct mg_object *mg_read_kml(const char *kml, int len);
-struct mg_object *mg_read_gml2(const char *gml, int len);
-struct mg_object *mg_read_gml3(const char *gml, int len);
+struct mg_object *mg_read_wkt(const char *wkt, size_t len);
+struct mg_object *mg_read_wkb(const char *wkb, size_t len);
+struct mg_object *mg_read_ewkt(const char *ewkt, size_t len);
+struct mg_object *mg_read_ewkb(const char *ewkb, size_t len);
+struct mg_object *mg_read_geojson(const char *json, size_t len);
+struct mg_object *mg_read_kml(const char *kml, size_t len);
+struct mg_object *mg_read_gml2(const char *gml, size_t len);
+struct mg_object *mg_read_gml3(const char *gml, size_t len);
 
-int mg_write_wkt(const struct mg_object *obj, char **wkt, int *len);
-int mg_write_wkb(const struct mg_object *obj, char **wkb, int *len);
-int mg_write_ewkt(const struct mg_object *obj, char **ewkt, int *len);
-int mg_write_ewkb(const struct mg_object *obj, char **ewkb, int *len);
-int mg_write_geojson(const struct mg_object *obj, char **json, int *len);
-int mg_write_kml(const struct mg_object *obj, char **kml, int *len);
-int mg_write_gml2(const struct mg_object *obj, char **gml, int *len);
-int mg_write_gml3(const struct mg_object *obj, char **gml, int *len);
+int mg_write_wkt(const struct mg_object *obj, char **wkt, size_t *len);
+int mg_write_wkb(const struct mg_object *obj, char **wkb, size_t *len);
+int mg_write_ewkt(const struct mg_object *obj, char **ewkt, size_t *len);
+int mg_write_ewkb(const struct mg_object *obj, char **ewkb, size_t *len);
+int mg_write_geojson(const struct mg_object *obj, char **json, size_t *len);
+int mg_write_kml(const struct mg_object *obj, char **kml, size_t *len);
+int mg_write_gml2(const struct mg_object *obj, char **gml, size_t *len);
+int mg_write_gml3(const struct mg_object *obj, char **gml, size_t *len);
 
 /* ------------------- geometry algorithm extern function ------------------- */
